+++ packages/api/nestjs-typeorm/lib/extensions/repository.ts
@@ -1,4 +1,4 @@
-import { And, type EntityManager, type EntityTarget, Equal, FindOneOptions, FindOperator, FindOptionsOrder, FindOptionsSelect, FindOptionsWhere, LessThan, MoreThan, ObjectLiteral, Repository } from 'typeorm'
+import { And, DeepPartial, type EntityManager, type EntityTarget, Equal, FindOneOptions, FindOperator, FindOptionsOrder, FindOptionsSelect, FindOptionsWhere, LessThan, MoreThan, ObjectLiteral, ObjectType, QueryDeepPartialEntity, Repository } from 'typeorm'
 import { createTransactionManagerProxy } from './transaction.js'
 import { createReadonlyManagerProxy } from './readonly.js'
 
@@ -7,6 +7,37 @@ export class TypeOrmRepository<T extends ObjectLiteral> extends Repository <T> {
     const proxy = createTransactionManagerProxy(createReadonlyManagerProxy(manager))
 
     super(entity, proxy)
+  }
+
+  async createAndInsert (entityLike: DeepPartial<T> | T): Promise<T>
+  async createAndInsert (entityLike: Array<DeepPartial<T> | T>): Promise<T[]>
+  async createAndInsert (entityLike: DeepPartial<T> | T | Array<DeepPartial<T> | T>): Promise<T | T[]> {
+    if (Array.isArray(entityLike)) {
+      const EntityClass = this.target as ObjectType<T>
+
+      const entities = entityLike.map(item =>
+        item instanceof EntityClass ? item as T : this.create(item as DeepPartial<T>)
+      )
+
+      await this.insert(entities as QueryDeepPartialEntity<T>[])
+
+      return entities
+    }
+
+    const EntityClass = this.target as ObjectType<T>
+    const isEntity = entityLike instanceof EntityClass
+
+    if (isEntity) {
+      await this.insert(entityLike as QueryDeepPartialEntity<T>)
+
+      return entityLike as T
+    } else {
+      const entity = this.create(entityLike)
+
+      await this.insert(entity as QueryDeepPartialEntity<T>)
+
+      return entity
+    }
   }
 
   async findNextBatch (
@@ -116,16 +147,16 @@ export class TypeOrmRepository<T extends ObjectLiteral> extends Repository <T> {
     for (let i = keys.length - 1; i >= 0; i--) {
       const key = keys[i]
       const keyLastEntityValue = keysLastEntityValues[i]
-      const preceedingKeys = keys.slice(0, i)
-      const preceedingKeysLastEntityValues = keysLastEntityValues.slice(0, i)
+      const precedingKeys = keys.slice(0, i)
+      const precedingKeysLastEntityValues = keysLastEntityValues.slice(0, i)
 
-      const preceedingKeysWhere = Object.fromEntries(
-        preceedingKeys.map((k, i) => [k, preceedingKeysLastEntityValues[i]])
+      const precedingKeysWhere = Object.fromEntries(
+        precedingKeys.map((k, i) => [k, precedingKeysLastEntityValues[i]])
       )
 
       const clause = {
         ...where,
-        ...preceedingKeysWhere,
+        ...precedingKeysWhere,
         [key]: this.getKeyCondition(where, order, key, keyLastEntityValue)
       } as FindOptionsWhere<T>