+++ packages/web/design-system/src/ui/dashboard-page/detail-pane/detailPane.type.ts
@@ -9,6 +9,7 @@ export interface DetailPaneStorage {
 
 export interface DetailPaneConfig {
   isResizable?: boolean
+  isToggleHidden?: boolean
   storage?: DetailPaneStorage | null
   variant?: DetailPaneVariant
 }