+++ packages/web/design-system/src/ui/date-field/datePickerField.style.ts
@@ -50,8 +50,8 @@ export const createDatePickerFieldStyle = tv({
       group-data-selected/cell:bg-white
     `,
     trigger: `
-      flex size-full min-w-0 cursor-pointer items-center bg-transparent text-xs
-      text-primary outline-none
+      flex size-full min-w-0 items-center bg-transparent text-xs text-primary
+      outline-none
       disabled:cursor-not-allowed disabled:text-disabled
       data-placeholder:text-placeholder
     `,