+++ packages/web/design-system/src/ui/table/context/tableGroup.context.ts
@@ -1,0 +1,10 @@
+import { useContext } from '@/composables/context.composable'
+
+interface TableGroupContext {
+  isGroup: boolean
+}
+
+export const [
+  useProvideTableGroupContext,
+  useInjectTableGroupContext,
+] = useContext<TableGroupContext>('tableGroupContext')