+++ packages/web/design-system/src/ui/table/context/tableSubGroup.context.ts
@@ -1,0 +1,10 @@
+import { useContext } from '@/composables/context.composable'
+
+interface TableSubGroupContext {
+  isSubGroup: boolean
+}
+
+export const [
+  useProvideTableSubGroupContext,
+  useInjectTableSubGroupContext,
+] = useContext<TableSubGroupContext>('tableSubGroupContext')