+++ packages/api/coordinates/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @wisemen/coordinates
 
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wisemen/quantity@0.8.4
+
+## 0.0.30
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wisemen/quantity@0.8.3
+
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wisemen/quantity@0.8.2
+
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wisemen/quantity@0.8.1
+
 ## 0.0.27
 
 ### Patch Changes