+++ packages/api/monetary/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @wisemen/monetary
 
+## 0.4.7
+
+### Patch Changes
+
+- Updated dependencies [[`0bc5b42`](https://github.com/wisemen-digital/wisemen-core/commit/0bc5b42c8a9f15e696e01dc24b6fd228b6657405)]:
+  - @wisemen/nestjs-typeorm@1.1.1
+
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies [[`23b823f`](https://github.com/wisemen-digital/wisemen-core/commit/23b823fd570e61b7d368b782e5330f9fd372e15d)]:
+  - @wisemen/nestjs-typeorm@1.1.0
+
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies [[`0c248c2`](https://github.com/wisemen-digital/wisemen-core/commit/0c248c2b889d75496817e0fd5c8c5f217c1f9deb)]:
+  - @wisemen/nestjs-typeorm@1.0.2
+
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies [[`9f7b309`](https://github.com/wisemen-digital/wisemen-core/commit/9f7b3095e56284975d3b28e1ca1b4dfe25dc032f)]:
+  - @wisemen/nestjs-typeorm@1.0.1
+
 ## 0.4.3
 
 ### Patch Changes