+++ packages/api/nestjs-typeorm/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @wisemen/nestjs-typeorm
 
+## 1.1.1
+
+### Patch Changes
+
+- [#1190](https://github.com/wisemen-digital/wisemen-core/pull/1190) [`0bc5b42`](https://github.com/wisemen-digital/wisemen-core/commit/0bc5b42c8a9f15e696e01dc24b6fd228b6657405) Thanks [@Kobe-Kwanten](https://github.com/Kobe-Kwanten)! - feat(TBN-1200): add AndOrIgnore FindOperator
+
+## 1.1.0
+
+### Minor Changes
+
+- [`23b823f`](https://github.com/wisemen-digital/wisemen-core/commit/23b823fd570e61b7d368b782e5330f9fd372e15d) Thanks [@jorenvandeweyer](https://github.com/jorenvandeweyer)! - Add array support to `createAndInsert` — passing an array now returns `Promise<T[]>` and inserts all entities in a single call
+
+## 1.0.2
+
+### Patch Changes
+
+- [#1162](https://github.com/wisemen-digital/wisemen-core/pull/1162) [`0c248c2`](https://github.com/wisemen-digital/wisemen-core/commit/0c248c2b889d75496817e0fd5c8c5f217c1f9deb) Thanks [@jorenvandeweyer](https://github.com/jorenvandeweyer)! - Add `createAndInsert` method to TypeORM repository extension
+
+## 1.0.1
+
+### Patch Changes
+
+- [#1053](https://github.com/wisemen-digital/wisemen-core/pull/1053) [`9f7b309`](https://github.com/wisemen-digital/wisemen-core/commit/9f7b3095e56284975d3b28e1ca1b4dfe25dc032f) Thanks [@jorenvandeweyer](https://github.com/jorenvandeweyer)! - Add createAndInsert function on TypeOrm repository
+
 ## 1.0.0
 
 ### Major Changes