+++ packages/api/twilio/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @wisemen/twilio
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [[`5266dcc`](https://github.com/wisemen-digital/wisemen-core/commit/5266dcce66c1df54ba1fbccb831186800c8db66b)]:
+  - @wisemen/api-error@1.0.0
+
 ## 0.0.3
 
 ### Patch Changes