+++ packages/web/tailwind-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @wisemen/vue-core-tailwind-config
 
+## 0.0.4
+
+### Patch Changes
+
+- [#1182](https://github.com/wisemen-digital/wisemen-core/pull/1182) [`26cbf86`](https://github.com/wisemen-digital/wisemen-core/commit/26cbf8667670a50e95e2cc2698eab3b9d93eb604) Thanks [@wouterlms](https://github.com/wouterlms)! - Reduce hover state contrast
+
+## 0.0.3
+
+### Patch Changes
+
+- [#1123](https://github.com/wisemen-digital/wisemen-core/pull/1123) [`44f3a20`](https://github.com/wisemen-digital/wisemen-core/commit/44f3a20c37e721204b2f08b13b9835c57594b71f) Thanks [@wouterlms](https://github.com/wouterlms)! - Improvements
+
 ## 0.0.2
 
 ### Patch Changes